+++ workspaces/marketplace/plugins/marketplace-common/src/index.ts
@@ -21,5 +21,6 @@
  */
 
 export * from './types';
+export * from './utils';
 export * from './api';
 export * from './utils';