+++ workspaces/marketplace/plugins/marketplace-backend/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @red-hat-developer-hub/backstage-plugin-marketplace-backend
 
+## 0.0.7
+
+### Patch Changes
+
+- 9a3c185: replace custom aggregation api endpoint with catalog entity-facets api
+- Updated dependencies [9a3c185]
+  - @red-hat-developer-hub/backstage-plugin-marketplace-common@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [0bcb97f]
+  - @red-hat-developer-hub/backstage-plugin-marketplace-common@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- 381a250: Upgrade Backstage from 1.32.0 to 1.35.0
+- Updated dependencies [381a250]
+  - @red-hat-developer-hub/backstage-plugin-marketplace-common@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes